Rewrite hdf5 files
Previously hdf5 archives were opened in 'a' mode, which rewrote the
data but kept the old data on disk, which was (most probably)
inaccessible. This was not the case for all data -- some of it was
correctly rewritten.
In mode 'w', the archive is completely rewritten. This can destroy
an already existing file, so we check that before the simulation.