Change v to positive. Change integration domain to tmax_gf0
quasimc.cpp test was also added.

v-domain coordinates were change to positive. Coordinates are now:
    -t_start < u_n < ... < u_1 < 0
    v_i = u_i - u_i+1 (before u_i+1 - u_i)
In a lot of (test) files, the model had to change from x to -x (x is v)

Now it is possible to set `keep_cubic_domain` to False. With this
setting, only points outside tmax_gf0 are rejected.

`do_importance_sampling` parameter was removed since it is not needed
any more (because of older commits).