Fix bug in measure
Commit 573b9490802ce60b7a4670de56be7e12200fe4ee has added a bug affecting
the kernel method in importance_sampling.cpp and measure.cpp.

We have to normalize the kernels -- divide them with model.weight. However,
in TwoDetKernelMeasure::accumulate they are also normalized with
config.accepted_weight, which is a remainder of the Quantum MCMC code.
Clearly, config.accepted_weight should be model.weight.

On the other hand, we want to keep track of the integrated integrand's
weight, which is config.accepted_weight/mode.weight.
Clearly, config.accepted_weight should be DIVIDED with model.weight.

To solve the problem, I added config.model_weight, which is now used
in TwoDetKernelMeasure::accumulate.
For Quantum MCMC, config.model_weight = config.accepted_weight.